requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
frozenlist==1.5.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-tools==1.68.1
idna==3.10
jwt==1.3.1
multidict==6.1.0
paypalrestsdk==1.13.3
propcache==0.2.1
protobuf==5.29.1
psycopg2-binary==2.9.10
pycparser==2.22
pyOpenSSL==24.3.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
